Revert "[BD-34] New dev docs outline and user interface (#35)"
This reverts commit 6279918b

Although we would like to continue with these
changes in the future, it was incomplete (i.e.
made use of stubbed documentation files), and
this project is temporarily on hold. Therefore,
this revert is to put things back as they were
until we have the resources to see this effort
through.

Changes there were not reverted:
1. Some conf updates.
2. Requirements updates.
3. Custom JS for externatl links.

ARCHBOM-1698