docs: rename developers_guide to micro_frontends (#56)
* docs: rename developers_guide to micro_frontends

This is part one of a proposed re-organization of this repository.

This renames the "developers_guide" directory to "micro_frontends" and simplifies the content to be specific to MFEs (with no meaningful content lost).  That's all that was in there anyway.

Other topics will get other sibling sub-directories, with the expectation that we may have "Developer's Guide"s for particular topics, but that the whole repository is "Developer Documentation".  Having both "Developer's Guide" and "Developer Documentation" as top-level concepts feels confusing.

Specifically, I'm hoping to move the content in  edx-documentation/en_us/developers into this repository as siblings to the "micro_frontends" content.  TBD the exact organization/information architecture, but this commit feels like a simple, non-controversial start.

* docs: adding a redirect to the new MFE content.

* docs: Apply suggestions from code review

Use `:doc:` links to link within the book.

Co-authored-by: Ned Batchelder <ned@nedbatchelder.com>

Co-authored-by: Feanil Patel <feanil@edx.org>
Co-authored-by: Ned Batchelder <ned@nedbatchelder.com>