chore: Upgrade python requirements.
We got in a situation where the newest version of pip breaks the pinned
version of pip-tools from running correctly.

https://github.com/jazzband/pip-tools/issues/1558

The issue is fixed in newer version of pip-tools so the only way to
resolve this is to first manually upgrade pip-tools and re-compile the
pip-tools.txt requirements file and then run `make upgrade`.