docs: Add the redirect extension.
Most of the content that lives here actually lives in a more up-to-date
form in some other place. Add the redirect plugin so that we can start
replacing content in this file with its equivalent in other places.

This is in preparation for when we can stop publishing this set of docs
all together.